0.0.1 api_err_header fixed
